pandicMdFlavour changed to choosePandocMdFlavaour
Done in belief that problem wi the colon at start of :pandocMdFlavour:gfm: was a problem from being like a keyword :pandoc
BUT in reality it is probably due to chooseRrender containing :
```if ( text.includes(":choosePandocMdFlavour:" ) ) {
    console.log( "contains :choosePandocMdFlavour:" );
    console.log( `text = \n ${text}`)```
    return renderPandocMdFlavour( text, filePath );
thereafter there is another test for flavour in renderPandocMdFlavour.
SO put test in a subroutine checkMdFlavour which defines & checks validity of flavour which is called by chooseRender to decide whether to call renderPandocMdFlavour if matched or bypass shis and continue with markdown-review md render.
The next commit will create